simplify tests and remove duplication. Add fixtures
