+++ tests/db/adapters/sqlite/test_feed_post_adapter.py
@@ -1,34 +1,19 @@
 """Tests for db.adapters.sqlite.feed_post_adapter module."""
 
 import sqlite3
-from contextlib import contextmanager
-from unittest.mock import MagicMock, Mock, patch
+from unittest.mock import Mock
 
 import pytest
 
 from db.adapters.sqlite.feed_post_adapter import SQLiteFeedPostAdapter
 from simulation.core.models.posts import BlueskyFeedPost
+from tests.db.adapters.sqlite.conftest import create_mock_db_connection, create_mock_row
 
 
 @pytest.fixture
 def adapter():
     """Create a SQLiteFeedPostAdapter instance."""
     return SQLiteFeedPostAdapter()
-
-
-def create_mock_row(row_data: dict) -> MagicMock:
-    """Helper function to create a mock sqlite3.Row.
-
-    Args:
-        row_data: Dictionary mapping column names to values
-
-    Returns:
-        MagicMock configured to behave like a sqlite3.Row
-    """
-    mock_row = MagicMock()
-    mock_row.__getitem__ = Mock(side_effect=lambda key: row_data[key])
-    mock_row.keys = Mock(return_value=list(row_data.keys()))
-    return mock_row
 
 
 @pytest.fixture
@@ -40,23 +25,9 @@ def mock_db_connection():
             mock_cursor.fetchall = Mock(return_value=[row1, row2])
             # test code here
     """
-
-    @contextmanager
-    def _mock_db_connection():
-        # Patch where it's used, not where it's defined
-        # This is necessary because get_connection is imported at module level
-        with patch(
-            "db.adapters.sqlite.feed_post_adapter.get_connection"
-        ) as mock_get_conn:
-            mock_conn = MagicMock()
-            mock_cursor = MagicMock()
-            mock_conn.__enter__ = Mock(return_value=mock_conn)
-            mock_conn.__exit__ = Mock(return_value=None)
-            mock_conn.execute.return_value = mock_cursor
-            mock_get_conn.return_value = mock_conn
-            yield mock_get_conn, mock_conn, mock_cursor
-
-    return _mock_db_connection
+    return create_mock_db_connection(
+        "db.adapters.sqlite.feed_post_adapter.get_connection"
+    )
 
 
 class TestSQLiteFeedPostAdapterReadFeedPostsByUris:
